+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Detection of syntax error from Redis response
 - Allow multiple level of fuzziness
+- Escape values in the query builder
 
 ## [1.1.0]